% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{bind_for}
\alias{bind_for}
\title{List Rendering}
\usage{
bind_for(tag, var, items)
}
\arguments{
\item{tag}{An htmltools \verb{<template>} tag.}

\item{var}{The loop variable name (e.g. \code{"item"}).}

\item{items}{A JavaScript expression for the list
(e.g. \code{"model.items"}).}
}
\value{
The modified tag.
}
\description{
Iterates over a list in the model. Must be used on a \verb{<template>}
tag. Inside the template, use the loop variable name to refer to
each item (e.g. \code{"item.name"}).
}
\examples{
shiny::tags$template(
  shiny::tags$li() |> bind_text("item")
) |>
  bind_for("item", "model.items")

}
