% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{on_change}
\alias{on_change}
\title{Change Event Handler (Static Value)}
\usage{
on_change(tag, msg, value)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{msg}{Character string specifying the message type.}

\item{value}{A static R value sent with the message.}
}
\value{
The modified tag.
}
\description{
Dispatches a message with a static value on change. Useful for
checkbox group items where each checkbox sends a fixed value.
}
\examples{
shiny::tags$input(type = "checkbox", value = "cheese") |>
  on_change("toggle_topping", "cheese")

}
